# kv: refactor and optimize