Optimize KV interface and implementation to support more efficient reads/writes, especially for remote backends.

bolt2 also receives performance boost because of this change (15-60%):
```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkNamePredicate-8                      118783        143410        +20.73%
BenchmarkLargeSetsNoIntersection-8            11655012      7952428       -31.77%
BenchmarkVeryLargeSetsSmallIntersection-8     21012045      16882451      -19.65%
BenchmarkHelplessContainsChecker-8            668915823     284239893     -57.51%
BenchmarkHelplessNotContainsFilms-8           668367233     285839315     -57.23%
BenchmarkHelplessNotContainsActors-8          673131537     287670777     -57.26%
BenchmarkNetAndSpeed-8                        1100455       930823        -15.41%
BenchmarkKeanuAndNet-8                        602669        578779        -3.96%
BenchmarkKeanuAndSpeed-8                      922366        788280        -14.54%
BenchmarkKeanuOther-8                         5592603       4209937       -24.72%
BenchmarkKeanuBullockOther-8                  9871332       7874863       -20.22%
BenchmarkSaveBogartPerformances-8             304158        295544        -2.83%

benchmark                                     old allocs     new allocs     delta
BenchmarkNamePredicate-8                      485            513            +5.77%
BenchmarkLargeSetsNoIntersection-8            79726          52505          -34.14%
BenchmarkVeryLargeSetsSmallIntersection-8     142217         116772         -17.89%
BenchmarkHelplessContainsChecker-8            5428455        2057189        -62.10%
BenchmarkHelplessNotContainsFilms-8           5434469        2060715        -62.08%
BenchmarkHelplessNotContainsActors-8          5431560        2060496        -62.06%
BenchmarkNetAndSpeed-8                        7267           6524           -10.22%
BenchmarkKeanuAndNet-8                        4075           4060           -0.37%
BenchmarkKeanuAndSpeed-8                      6011           5418           -9.87%
BenchmarkKeanuOther-8                         37148          28334          -23.73%
BenchmarkKeanuBullockOther-8                  62192          47335          -23.89%
BenchmarkSaveBogartPerformances-8             1730           1689           -2.37%

benchmark                                     old bytes     new bytes     delta
BenchmarkNamePredicate-8                      27579         29033         +5.27%
BenchmarkLargeSetsNoIntersection-8            4520475       3107477       -31.26%
BenchmarkVeryLargeSetsSmallIntersection-8     8044333       6675525       -17.02%
BenchmarkHelplessContainsChecker-8            298876692     128403280     -57.04%
BenchmarkHelplessNotContainsFilms-8           299226556     128626841     -57.01%
BenchmarkHelplessNotContainsActors-8          298952092     128490849     -57.02%
BenchmarkNetAndSpeed-8                        444931        407117        -8.50%
BenchmarkKeanuAndNet-8                        251183        251374        +0.08%
BenchmarkKeanuAndSpeed-8                      364451        333921        -8.38%
BenchmarkKeanuOther-8                         2210128       1749142       -20.86%
BenchmarkKeanuBullockOther-8                  3622010       2843176       -21.50%
BenchmarkSaveBogartPerformances-8             97451         94658         -2.87%
```