@jtorvald 2017-09-28
Impressive.